PR3: Complete - split ConvergenceCriteria and Simulators into submodules
Split ConvergenceCriteria.py into Simulation/Convergence/:
- base.py: ConvergenceCriterion ABC + _is_momentum_solver_instance helper
- residuals.py: shared numeric helpers (_compute_force_residual, etc.)
- strain_based.py: StrainBasedCriterion
- force_residual.py: ForceResidualCriterion
- displacement_increment.py: DisplacementIncrementCriterion + _initialize_step_displacement
- force_displacement.py: ForceDisplacementCriterion (composite)
- handler.py: ConvergenceErrorHandler + resolve_convergence_criterion + factory functions
- __init__.py: re-exports all public classes/functions

Split Simulators.py into Simulation/Simulators/:
- base.py: Simulator ABC + all snapshot/restore/bisection scaffolding
- thermo_mechanical.py: Simulator_TM (coupled thermo-mechanical)
- mechanical.py: Simulator_M (mechanical-only, staggered)
- thermal.py: Simulator_T (thermal-only)
- mechanical_out.py: Simulator_Mout (deprecated wrapper, un-exported)
- __init__.py: re-exports all public simulator classes

Converted old flat files to shims:
- ConvergenceCriteria.py → re-export shim
- Simulators.py → re-export shim

Fixed type annotations:
- Used Any for runtime type hints to avoid NameError from TYPE_CHECKING-only imports
- Added __future__ annotations for Python 3.10+ union syntax (str | Type)
- Added necessary module-level imports to all submodule files

Verification:
- Clean import (no cycles or circular dependencies)
- Old paths work (from safeincave.MomentumEquation import LinearMomentum)
- New paths work (from safeincave.Equations.Momentum.standard import LinearMomentum)
- Backward compatibility maintained for Simulator_Mout (not exported at top level)

Next: PR4 (constitutive move + discovery upgrade) and PR5 (new solver implementation)
